Move `prompt` functions to `prompt.py`
