+++ operate/prompts/prompt.py
@@ -1,3 +1,8 @@
+from operate.config.settings import Config
+
+config = Config()
+monitor_size = config.monitor_size
+
 # General user Prompts
 USER_QUESTION = "Hello, I can help you with anything. What would you like done?"
 
@@ -191,3 +196,35 @@ Original objective: {objective}
 
 Display the results clearly:
 """
+
+
+def format_summary_prompt(objective):
+    """
+    Format the summary prompt
+    """
+    prompt = SUMMARY_PROMPT.format(objective=objective)
+    return prompt
+
+
+def format_vision_prompt(objective, previous_action):
+    """
+    Format the vision prompt
+    """
+    if previous_action:
+        previous_action = f"Here was the previous action you took: {previous_action}"
+    else:
+        previous_action = ""
+    prompt = VISION_PROMPT.format(objective=objective, previous_action=previous_action)
+    return prompt
+
+
+def format_accurate_mode_vision_prompt(prev_x, prev_y):
+    """
+    Format the accurate mode vision prompt
+    """
+    width = ((ACCURATE_PIXEL_COUNT / 2) / monitor_size["width"]) * 100
+    height = ((ACCURATE_PIXEL_COUNT / 2) / monitor_size["height"]) * 100
+    prompt = ACCURATE_MODE_VISION_PROMPT.format(
+        prev_x=prev_x, prev_y=prev_y, width=width, height=height
+    )
+    return prompt
